Add levels, make GameDead function, remove a few of warning.
StringSprite.cpp - remove a warning that implicit casting size_t to int is resolved via static_cast<>.
CameraManager.cpp&.hpp - add Camera initialize function that will be used at GameDead event. Renaming Graphics::CameraManager::ZoomAndCollisionBoxHandling into Graphics::CameraManager::ZoomAndCollisionRegionHandling.
ObjectManager.cpp - fix a small data type casting issue.
String.cpp&hpp - Make an InitString function that initialize all string vertex, and from now on string store an each pointer to player data.
BasicMovementLevel.cpp&hpp, OneWayPassLevel.cpp&hpp, SizeScalingLevel.cpp&hpp - Add a file of level.
State.cpp - implement draw function in here (base class)
State.hpp - Improve ctor and CameraManager move into here.
TestLevel.cpp&hpp - Add uniform Init list at member variable. Move draw function from here into State base class. Add GameDead function that will be called when player is dead.

CIB - No one...

+++ Gam200Engine/Component/StringSprite.cpp
@@ -66,7 +66,7 @@ void StringSprite::UpdateUniforms(const matrix3& toNDC, float depth)
 
 	// String uniforms upgrade
 	size_t verticesSize = stringOwner->vertices.size();
-	material->intUniform[Graphics::SHADER::Uniform_String_Vertex_Capacity] = verticesSize;
+	material->intUniform[Graphics::SHADER::Uniform_String_Vertex_Capacity] = static_cast<int>(verticesSize);
 	material->floatUniforms[Graphics::SHADER::Uniform_String_Height] = stringHeight;
 
 	// Initialize static array of vector2
